001 - Introduce enums for consistent data representation
BookmakerType - Statically fills out all the Bookmakers.
A map is created so that a BookmakerKey can be used to find the
relevant BookmakerType in constant time.

MarketType - Different Types of markets that we are currently
interested in.

Region - The different regions each bookie belongs to.

Sport - The type of sport.