030 - Handle Exceptions that are recoverable
Some exceptions such as SchemaMappingExceptions are acceptable and the
program should be able to be continue on, at this point of the program.

Exceptions should go silent and caller should have the ability to handle
them how they see fit, while not stopping the program.

It is extendible to be able to create different kinds of
Exceptionhandlers such as, No-Op exception handlers, or handlers that
prompt message to the GUI.