039 - Add methods to use MetadataCache in FeedService
MetadataCache is used to initialize a trader agent, and the source of
the most updated state the system is aware of so far.

It gets updated with ONLY the data that it will publish to its
subscribers.

It will compare elements to see if data is a new event to be published
or not.