042 - Create TradeAgents which use TradingStrategy to send TradeAction
TradeAgents all have their own view of the book that they are allowed to
trade according to their config.

During initialization, all permutations of BookmakerGroupingKey is
created, despite a Region or Market not being available at the time.

TradeAgents must evaluate all incoming feed to see if it can create a
TradeAction.

We do not care about `MarketToBookmakersUpdateEvent` because we will
lazy fill new Bookmakers with `OddsMetadataUpdate`.

We could do it, however, we would need to put a place holder odd or
null, which is a SMELL.